Move User and Tenant entity to own database, left all tenants in one database for the moment
