+++ .env
@@ -26,7 +26,8 @@ APP_SECRET=ec106247aa347f0d4435c79ea55f5136
 # DATABASE_URL="sqlite:///%kernel.project_dir%/var/data.db"
 # DATABASE_URL="mysql://app:!ChangeMe!@127.0.0.1:3306/app?serverVersion=8.0.32&charset=utf8mb4"
 # DATABASE_URL="mysql://app:!ChangeMe!@127.0.0.1:3306/app?serverVersion=10.11.2-MariaDB&charset=utf8mb4"
-DATABASE_URL="postgresql://app:!ChangeMe!@127.0.0.1:5432/app?serverVersion=15&charset=utf8"
+DATABASE_LANDLORD_URL="postgresql://app:!ChangeMe!@127.0.0.1:5432/landlord?serverVersion=15&charset=utf8"
+DATABASE_TENANT_URL="postgresql://app:!ChangeMe!@127.0.0.1:5432/tenant?serverVersion=15&charset=utf8"
 ###< doctrine/doctrine-bundle ###
 
 SERVER_NAME="tenant1.localhost, tenant2.localhost, tenant3.localhost, tenant1.app.localhost, tenant2.app.localhost, caddy:80"
